+++ frontend/.env
@@ -1,2 +1,4 @@
-PUBLIC_POCKETBASE_URL="https://chat.christopherbaklid.com/pocketbase"
-PUBLIC_RELAY_URL="https://chat.christopherbaklid.com/relay"
+#PUBLIC_POCKETBASE_URL="https://chat.christopherbaklid.com/pocketbase"
+#PUBLIC_RELAY_URL="https://chat.christopherbaklid.com/relay"
+PUBLIC_POCKETBASE_URL="http://frontend.docker.localhost/pocketbase"
+PUBLIC_RELAY_URL="http://frontend.docker.localhost/relay"
