--- frontend/src/components/Menu.svelte
@@ -1,24 +1,0 @@
-<script>
-	import { ArrowRightFromBracketSolid, CogOutline } from 'flowbite-svelte-icons';
-</script>
-
-<div class="bg-neutral w-40">
-	<ul class="menu menu-horizontal">
-		<li>
-			<!-- svelte-ignore a11y-missing-attribute -->
-			<a>
-				<div class="h-5 w-5">
-					<ArrowRightFromBracketSolid />
-				</div>
-			</a>
-		</li>
-		<li>
-			<!-- svelte-ignore a11y-missing-attribute -->
-			<a>
-				<div class="h-5 w-5">
-					<CogOutline />
-				</div>
-			</a>
-		</li>
-	</ul>
-</div>
