+++ frontend/src/components/Navbar.svelte
@@ -1,12 +1,6 @@
 <script>
-	import { currentUser, pb } from '$lib/pocketbase';
-	import { Disconnect } from '../lib/grpc';
 	import { server } from '../stores/server';
-
-	function logout() {
-		pb.authStore.clear();
-		Disconnect();
-	}
+	import Login from './Login.svelte';
 </script>
 
 <div class="navbar bg-base-300">
@@ -15,5 +9,8 @@
 			<div class="w-2 rounded-full bg-green-400" />
 		</div>
 		<div class="btn btn-ghost upper-case text-xl">{$server}</div>
+	</div>
+	<div class="justify-end">
+		<Login />
 	</div>
 </div>
